+++ pyWebmConverter/__init__.py
@@ -2,7 +2,7 @@
 pyWebmConverter - A Python GUI application for converting video files to WebM format.
 """
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 __author__ = "pyWebmConverter Contributors"
 __license__ = "Unlicense"
 
@@ -10,6 +10,9 @@ __all__ = [
     "command_builder",
     "audio_processor",
     "ffmpeg_worker",
+    "ffmpeg_gui",
     "video_editor",
+    "waveform",
+    "notifier",
     "constants",
 ]